fix: 检查更新界面失去焦点问题
1.在检查更新和更新设置页签切换的过程中设置焦点
2.将下载大小计算的数据类型设置成int64,防止下载量过大计算出错

Log: 修复了测试版本有更新，更新设置界面，开启仅安全性更新，返回检查更新，全部更新按钮被选中的问题.
Bug: https://pms.uniontech.com/bug-view-120943.html
Influence: 更新展示

+++ src/frame/window/modules/update/updatectrlwidget.cpp
@@ -707,7 +707,7 @@ void UpdateCtrlWidget::onRequestRefreshSize()
     if (m_updateSize == 0) {
         m_CheckAgainBtn->setEnabled(false);
         m_upgradeWarningGroup->setVisible(false);
-    } else if ((static_cast<int>(m_updateSize) / 1024) / 1024 >= m_qsettings->value("upgrade_waring_size", UpgradeWarningSize).toInt()) {
+    } else if ((static_cast<int64_t>(m_updateSize) / 1024) / 1024 >= m_qsettings->value("upgrade_waring_size", UpgradeWarningSize).toInt()) {
         m_upgradeWarningGroup->setVisible(true);
     } else {
         m_upgradeWarningGroup->setVisible(false);
